// app-vehicle-manager/src/app/features/vehicles/model/vehicle.model.ts
export interface VehicleModel {
  idVehiculo?: string;
  placa?: string;
  anio?: string;
  color?: string;
  kilometraje?: string;
  vin?: string;
  tipoCombustible?: string;
  transmision?: string;
  estadoVehiculo?: string;
  idModelo?: string;
  modeloNombre?: string;
  idMarca?: string;
  marcaNombre?: string;
  precio?: string;
}